ajustes

+++ src/components/CourseForm.jsx
@@ -27,17 +27,24 @@ export default function CourseForm() {
                 if (res.status === 201) {
                     return alert(res.data)
                 }
+
+                else if (res.status === 400) {
+                    return alert(res.data)
+                }
                 
                 else if (res.status === 401) {
-                    localStorage.clear()
-                    
+                    localStorage.clear()                  
                     return router.replace("/")
                 }
 
                 return console.log(res.data)
             })
             .catch((err) => {
-                if (err.response.status === 401) {
+                if (err.response.status === 400) {
+                    return alert(err.response.data)
+                }
+
+                else if (err.response.status === 401) {
                     localStorage.clear()
                     return router.replace("/")
                 }
